requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
djangorestframework==3.15.2
idna==3.10
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0

pip~=24.0
pytz~=2024.2
PyYAML~=6.0.2
packaging~=24.2
inflection~=0.5.1
uritemplate~=4.1.1
setuptools~=65.5.0
PyJWT~=2.10.1
django-filter~=24.3